port markdown-inline scanner to c
